removed key and language from SQLDatabase::prepare
Since this "feature" has never really ever gotten any use and only serves
to add a lot of clutter in code, the specification has had it removed.
It's recomended to implement the feature as methods but the spec will try
to keep compatibility with the ($key, $statemant, $lang) signature going
forward.